requirements.txt
streamlit
yfinance
pandas
google-generativeai
matplotlib
pandas-ta
